+++ index.html
@@ -1,0 +1,103 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <title>Interpretor</title>
+    <style>
+        .container {
+            max-width: 800px;
+            margin: 0 auto;
+            padding: 20px;
+        }
+        .form-group {
+            margin-bottom: 15px;
+        }
+        textarea {
+            width: 100%;
+            min-height: 200px;
+        }
+        select {
+            padding: 5px;
+            margin: 5px;
+        }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <h1>Interpretor</h1>
+        <div class="form-group">
+            <select id="fromLanguage">
+                <option value="English">English</option>
+                <option value="German">German</option>
+                <option value="Spanish">Spanish</option>
+                <option value="French">French</option>
+            </select>
+            <span>to</span>
+            <select id="toLanguage">
+                <option value="German">German</option>
+                <option value="English">English</option>
+                <option value="Spanish">Spanish</option>
+                <option value="French">French</option>
+            </select>
+        </div>
+        <div class="form-group">
+            <select id="tone">
+                <option value="direct">Direct Translation</option>
+                <option value="friendly">Friendly Manner</option>
+                <option value="formal">Formal Tone</option>
+                <option value="casual">Casual Tone</option>
+            </select>
+        </div>
+        <div class="form-group">
+            <textarea id="inputText" placeholder="Enter text to translate..."></textarea>
+        </div>
+        <button onclick="translate()">Translate</button>
+        <div class="form-group">
+            <h3>Translation:</h3>
+            <div id="outputText"></div>
+        </div>
+    </div>
+
+    <script>
+        async function translate() {
+            const fromLang = document.getElementById('fromLanguage').value;
+            const toLang = document.getElementById('toLanguage').value;
+            const tone = document.getElementById('tone').value;
+            const text = document.getElementById('inputText').value;
+
+            const systemPrompt = tone === 'friendly' 
+                ? "You are a friendly translator. Provide translations in a warm, conversational tone."
+                : "You are a direct translator. Provide only the translated words without any explanation or additional context.";
+
+            const requestBody = {
+                messages: [
+                    {
+                        role: "system",
+                        content: systemPrompt
+                    },
+                    {
+                        role: "user",
+                        content: `Translate from ${fromLang} to ${toLang}:\n${text}`
+                    }
+                ],
+                model: "mixtral-8x7b-32768"
+            };
+
+            try {
+                const response = await fetch('/translate', {
+                    method: 'POST',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    },
+                    body: JSON.stringify(requestBody)
+                });
+
+                const data = await response.json();
+                document.getElementById('outputText').textContent = data.translation;
+            } catch (error) {
+                console.error('Translation error:', error);
+                document.getElementById('outputText').textContent = 'Translation failed. Please try again.';
+            }
+        }
+    </script>
+</body>
+</html>
