+++ internal/models/chat.go
@@ -1,0 +1,17 @@
+package models
+
+type Message struct {
+	Role    string `json:"role"`
+	Content string `json:"content"`
+}
+
+type ChatRequest struct {
+	Messages []Message `json:"messages"`
+	Model    string    `json:"model"`
+}
+
+type ChatResponse struct {
+	Choices []struct {
+		Message Message `json:"message"`
+	} `json:"choices"`
+}
